refactor(store): ingredients, order slices
