+++ src/services/ingredients/initial-state.js
@@ -1,4 +1,4 @@
 export const initialState = {
-  status: undefined,
+  status: null,
   ingredients: [],
 };
