+++ src/services/order/reducers/clear-order.js
@@ -1,3 +1,3 @@
-import { initialState } from "../initialState";
+import { initialState } from "../initial-state";
 
 export const clearOrder = () => initialState;
