+++ src/services/order/reducers/index.js
@@ -1,1 +1,2 @@
 export { clearOrder } from "./clear-order";
+export { resetOrderIdStatus } from "./reset-order-id-status";
