+++ src/services/order/reducers/reset-order-id-status.js
@@ -1,0 +1,5 @@
+import { initialState } from "../initial-state";
+
+export const resetOrderIdStatus = (state) => {
+  state.orderIdStatus = initialState.orderIdStatus;
+};
